cd Fix (#50)
Fixed a redundancy on `free_ms()` and a mistake on `pwd()`.

Tried to make `cd()` properly change the environment variables values. However, there was an error when changing the environment variables in `export()`: `strcmp_chr()` was only comparing the character right before '=', instead of the whole string.

Had to change the way `strcmp_chr()` worked in order to fix `cd` and `export`, however that broke `unset`. This was due to the fact that `export` should compare the string up-to '=', including '=' if it exists, while `unset` should compare the string up-to '=' without including it.

Had to create a new function for `unset` to compare the necessary strings, since it wasn't supposed to include '=', but it still had to return "different" if it didn't reach the end of `s1`. I called the function `strcmp_nochr()`.

While testing my work I noticed another issue though: `cd` was segfaulting  when it received no arguments. This was apparently due to the fact that I was using `env_var_str()`. I noticed we were using `path_str()` as `env | grep PATH`, and since I wanted to basically do `env | grep HOME` I changed the functions name to `var_str()` and made it work with any environment variable it receives.

After all these changes, cd functions properly now.