Beggining of a prototype
I've only prototyped the redirections, still need to actually figure out the executer (forks, pipelines, etc)

Also `ft_exit` is wrong, I had changed it so I could get the exit status, but it was a shitty work-around. Since we still need to figure out how to get the exit status from the children, my work-around is gonna be irrelevant and we're gonna have to re-write `ft_exit` back to how it was before.